Integrate queue_check into redis_manager
