+++ tests/mirrulations_server_tests/test_redis_manager.py
@@ -1,13 +1,17 @@
 import fakeredis
 import json
 import mock
-from mirrulations_server.redis_manager import RedisManager
+from mirrulations_server.redis_manager import RedisManager, queue_check
 from ast import literal_eval
 import time
 
-@mock.patch('mirrulations_server.redis_manager.reset_lock')
-@mock.patch('mirrulations_server.redis_manager.set_lock')
-def make_database(reset, lock):
+
+def make_empty_database():
+    r = RedisManager(fakeredis.FakeRedis())
+    return r
+
+
+def make_unlocked_database():
     r = RedisManager(fakeredis.FakeRedis())
     r.delete_all()
     list = json.dumps({"A":"a", "B":["b", "c"]})
@@ -15,12 +19,26 @@ def make_database(reset, lock):
     list3 = json.dumps({"G":"g", "H":["h", "i"]})
     r.add_to_queue(list)
     r.add_to_queue(list2)
+    r.add_to_progress(list3)
+    return r
+
+
+@mock.patch('mirrulations_server.redis_manager.reset_lock')
+@mock.patch('mirrulations_server.redis_manager.set_lock')
+def make_locked_database(reset, lock):
+    r = RedisManager(fakeredis.FakeRedis())
+    r.delete_all()
+    list1 = json.dumps({"A":"a", "B":["b", "c"]})
+    list2 = json.dumps({"D":"d", "E":["e", "f"]})
+    list3 = json.dumps({"G":"g", "H":["h", "i"]})
+    r.add_to_queue(list1)
+    r.add_to_queue(list2)
     r.add_to_queue(list3)
     return r
 
 
 def ignore_test_iterate():
-    r = make_database()
+    r = make_locked_database()
     for item in r.get_all_items_in_queue():
         item = literal_eval(item.decode('utf-8'))
         if int(time.time()) - item[2] > 21600:
@@ -31,21 +49,35 @@ def ignore_test_iterate():
             print("Still Good!")
 
 
+def test_queue_check_empty():
+    r = make_empty_database()
+    a,b = queue_check(r)
+    assert len(a) == 0
+    assert len(b) == 0
+
+
+def test_queue_check_items_in_queue():
+    r = make_unlocked_database()
+    a,b = queue_check(r)
+    assert len(a) == 1
+    assert len(b) == 2
+
+
 def test_get_all_item_in_queue():
-    r = make_database()
+    r = make_locked_database()
     items = r.get_all_items_in_queue()
     assert len(items) == 3
 
 
 def test_get_all_item_in_progress():
-    r = make_database()
+    r = make_locked_database()
     list4 = json.dumps(["l", ["m", "n"]])
     r.add_to_progress(list4)
     assert len(r.get_all_items_in_progress()) == 1
 
 
 def test_get_work():
-    r = make_database()
+    r = make_locked_database()
     assert len(r.get_all_items_in_queue()) == 3
     assert len(r.get_all_items_in_progress()) == 0
     work = r.get_work()
@@ -55,7 +87,7 @@ def test_get_work():
 
 
 def test_add_to_queue():
-    r = make_database()
+    r = make_locked_database()
     list4 = json.dumps(["j", ["k", "l"]])
     assert len(r.get_all_items_in_queue()) == 3
     r.add_to_queue(list4)
@@ -63,7 +95,7 @@ def test_add_to_queue():
 
 
 def test_delete_all():
-    r = make_database()
+    r = make_locked_database()
     r.delete_all()
     assert len(r.get_all_items_in_queue()) == 0
     assert len(r.get_all_items_in_progress()) == 0
@@ -71,7 +103,7 @@ def test_delete_all():
 
 @mock.patch('mirrulations_server.redis_manager.get_curr_time', return_value=1531911498)
 def test_find_expired(time):#time):
-    r = make_database()
+    r = make_locked_database()
     r.delete_all()
     t3 = json.dumps((["g", ["h", "i"]]))
     r.add_to_progress(t3)
@@ -81,7 +113,7 @@ def test_find_expired(time):#time):
 
 
 def test_find_no_expired():
-    r = make_database()
+    r = make_locked_database()
     r.delete_all()
     t3 = json.dumps((["g", ["h", "i"]]))
     t2 = json.dumps((["j", ["k", "l"]]))
@@ -93,21 +125,21 @@ def test_find_no_expired():
 
 
 def test_get_specific_item_from_queue():
-    r = make_database()
+    r = make_locked_database()
     r.delete_all()
     r.add_to_queue(json.dumps({"A":"B", "job_id":"d"}))
     assert r.get_specific_job_from_queue("d") == json.dumps({"A": "B", "job_id": "d"})
 
 
 def test_get_specific_item_from_queue_does_not_match():
-    r = make_database()
+    r = make_locked_database()
     r.delete_all()
     r.add_to_queue(json.dumps({"A":"B", "job_id":"c"}))
     assert r.get_specific_job_from_queue("d") == '{"job_id":"null", "type":"none"}'
 
 
 def test_remove_specific_job_from_queue():
-    r = make_database()
+    r = make_locked_database()
     r.delete_all()
     r.add_to_queue(json.dumps({"A": "B", "job_id": "c"}))
     r.add_to_queue(json.dumps({"A": "B", "job_id": "d"}))
@@ -119,7 +151,7 @@ def test_remove_specific_job_from_queue():
 
 
 def test_remove_specific_job_from_queue_no_item():
-    r = make_database()
+    r = make_locked_database()
     r.delete_all()
     r.add_to_queue(json.dumps({"A": "B", "job_id": "c"}))
     r.add_to_queue(json.dumps({"A": "B", "job_id": "d"}))
@@ -130,7 +162,7 @@ def test_remove_specific_job_from_queue_no_item():
 
 @mock.patch('mirrulations_server.redis_manager.get_curr_time', return_value=15)
 def test_get_keys_progress(time):
-    r = make_database()
+    r = make_locked_database()
     r.delete_all()
     r.add_to_progress(json.dumps({"A": "B", "job_id": "c"}))
     r.add_to_progress(json.dumps({"A": "B", "job_id": "d"}))
@@ -139,7 +171,7 @@ def test_get_keys_progress(time):
 
 @mock.patch('mirrulations_server.redis_manager.get_curr_time', return_value=15)
 def test_does_job_exist_in_progress(time):
-    r = make_database()
+    r = make_locked_database()
     r.delete_all()
     r.add_to_progress(json.dumps({"A": "B", "job_id": "c"}))
     assert r.does_job_exist_in_progress("c")
@@ -147,7 +179,7 @@ def test_does_job_exist_in_progress(time):
 
 @mock.patch('mirrulations_server.redis_manager.get_curr_time', return_value=15)
 def test_delete_from_progress(time):
-    r = make_database()
+    r = make_locked_database()
     r.delete_all()
     r.add_to_progress(json.dumps({"A": "B", "job_id": "c"}))
     assert len(r.get_all_items_in_progress()) == 1
@@ -157,7 +189,7 @@ def test_delete_from_progress(time):
 
 @mock.patch('mirrulations_server.redis_manager.get_curr_time', return_value=15)
 def test_renew_job(time):
-    r = make_database()
+    r = make_locked_database()
     r.delete_all()
     r.add_to_progress(json.dumps({"A": "B", "job_id": "c"}))
     assert len(r.get_all_items_in_progress()) == 1
@@ -165,19 +197,3 @@ def test_renew_job(time):
     r.renew_job("c")
     assert len(r.get_all_items_in_queue()) == 1
     assert len(r.get_all_items_in_progress()) == 0
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
